% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AAPL.data}
\alias{AAPL.data}
\title{Apple (AAPL) stock prices and volumes from Yahoo Finance}
\format{An xts object with 1437 rows and 6 columns
\describe{
   \item{AAPL.Open}{Daily Opening Price}
   \item{AAPL.High}{Daily High}
   \item{AAPL.Low}{Daily Low}
   \item{AAPL.Volume}{Daily Volume}
   \item{AAPL.Adjusted}{Closing Price adjusted for splits and dividends}
}}
\source{
\url{http://finance.yahoo.com/}
}
\usage{
AAPL.data
}
\description{
Apple (AAPL) stock prices and volumes from Yahoo Finance
}
\keyword{datasets}

