% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_df.R
\name{as_ts_df}
\alias{as_ts_df}
\title{Converts a ts or mts object to a dataframe}
\usage{
as_ts_df(ts_obj, date_col = TRUE)
}
\arguments{
\item{ts_obj}{the ts or mts object to be converted into a dataframe}

\item{date_col}{Flag variable.  If TRUE, then the time variable in the new dataframe
will be a date object}
}
\value{
dataframe where a column of class "Date" where the first column is labeled "time"
}
\description{
\code{as.ts.df} transfroms a ts or mts object to a dataframe
}
\details{
The first column of the new dataframe holds the time/date in
in a column labeled "time"

Note: If the data are quarterly, the first quarter will be translated
to "YYYY-01-01", if monthly the first month will be translated to
"YYYY-01-01" and so on
}
\examples{
#USMacrsoSW form the AER package is an "mts" object
library(AER)
data("USMacroSW")
as_ts_df(USMacroSW)
#Using the AirPassengers data
data(AirPassengers)
as_ts_df(AirPassengers)

}
\seealso{
\code{\link{as_zoo_df}}
}

