% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_df.R
\name{as_zoo_df}
\alias{as_zoo_df}
\title{Convert a zoo or xts object into a dataframe}
\usage{
as_zoo_df(data.zoo)
}
\arguments{
\item{data.zoo}{zoo or xts object}
}
\value{
a data frame with the first column as a date object labeled "time"
}
\description{
Convert a zoo or xts object into a dataframe
}
\examples{
library(xts)
#The apple stock data is a xts object
data(AAPL.data)
class(AAPL.data)   # "xts" "zoo"
head(as_zoo_df(AAPL.data))

}
\seealso{
\code{\link{as_ts_df}}
}

