% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bear_dates}
\alias{bear_dates}
\title{Bear market dates}
\format{A data frame with 2 columns:
\describe{
   \item{start}{Start of bear market cycle}
   \item{end}{End of bear market cycle}
}}
\source{
Chandler Lutz's calculations
}
\usage{
bear_dates
}
\description{
Bear market dates defined as a 20 percent or more drop in the
S&P500 over a period of 2 or more months
}
\keyword{datasets}

