% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_cycle.R
\name{geom_cycle}
\alias{geom_cycle}
\title{A geom that draws vertical bars based on given dates. The default
dates are \code{recession_dates} corresponding to dates for NBER
recessions. The data also can use \code{bear_dates}}
\usage{
geom_cycle(dates = ggts::recession_dates, fill = "#003F87", alpha = 0.2,
  geom = "rect", position = "identity", show.legend = FALSE,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{dates}{(dataframe) a dataframe with the dates that will be
used to create the vertical bars. The first column is the start
of each cycle; the second column is the end of each
cycle. Default is \code{recession_dates}}

\item{fill}{(string) a string with the color used to fill the cycle
bars (corresponds to ggplot2 fill)}

\item{alpha}{(numeric) a number between 0 and 1 that specifies the
transparency of the regression bars (corresponds to ggplot2
alpha)}

\item{geom}{the ggplot2 geom to be plotted. Defaults to "rect"}

\item{position}{the ggplot2 position. Defaults to "identity"}

\item{show.legend}{Determines ggplot2 show legend
behavior. Defaults to FALSE}

\item{inherit.aes}{Speficies if the aes should be inherited from
previous plot. Defaults to TRUE}

\item{...}{Other parameters to be passed to the layer}
}
\description{
A geom that draws vertical bars based on given dates. The default
dates are \code{recession_dates} corresponding to dates for NBER
recessions. The data also can use \code{bear_dates}
}
\examples{
data(AirPassengers)
##The defualt shading for geom_cycle is recession dates
ggts(AirPassengers) + geom_cycle()

##Apple stock prices originally in \\code{xts} format
library(xts)
data(AAPL.data)  ##xts object
##Apple stock price with recessions shaded
ggts(AAPL.data, -dplyr::matches("Volume")) + geom_cycle()


##Apple stock price with recessions and bear markets shaded
##First, load the bear market dates and recession dates
data(bear_dates); data(recession_dates)
ggts(AAPL.data, -dplyr::matches("Volume")) + geom_cycle(dates = bear_dates)
ggts(AAPL.data, -dplyr::matches("Volume")) + geom_cycle(dates = recession_dates) +
   geom_cycle(dates = bear_dates, fill="gray50")




##Economics data
data(economics)
ggts(economics, psavert, uempmed) + geom_cycle()

##Using ggts_facet
ggts_facet(economics) + geom_cycle()
library(AER)
data(USMacroSW)
data(bear_dates)
ggts_facet(USMacroSW) + geom_cycle()
ggts_facet(USMacroSW) + geom_cycle(dates=bear_dates)
ggts_facet(USMacroSW) + geom_cycle() + geom_cycle(dates=bear_dates, fill="grey50")

##Can also be used with a ggplot not created by ggts
library(dplyr); library(tidyr)
econ.out <- economics \%>\%
   select(date, psavert, uempmed) \%>\%
   gather(variable, value, -date)
ggplot(econ.out, aes(x=date, y=value, group=variable)) +
   geom_line(aes(linetype=variable)) +
   geom_cycle()

}
\seealso{
\code{\link{ggts}}
}

