% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_cycle.R
\docType{data}
\name{GeomCycle}
\alias{GeomCycle}
\title{Create a stat that uses ggproto to
that will calculate the the bars to show for the
cycle}
\format{An object of class \code{GeomCycle} (inherits from \code{Stat}, \code{ggproto}) of length 3.}
\usage{
GeomCycle
}
\description{
Create a stat that uses ggproto to
that will calculate the the bars to show for the
cycle
}
\keyword{datasets}

