\docType{package}
\name{ggts-package}
\alias{ggts-package}
\title{Easy Time Series Visualization via gglot2}
\description{
  Easy Time Series Visualization via ggplot2
}
\details{
  \tabular{ll}{ Package: \tab ggts\cr Type: \tab Package\cr
  Version: \tab: 0.0.1.0\cr Date: \tab 2012-09-16\cr
  License: \tab GPL (>=2)\cr LazyLoad: \tab yes\cr }

  This package provides easy time series visualization via
  ggplot2. The ggts package plots ts, zoo, and xts objects;
  implements bear market, recession, or other shading; and
  allows for standard or faceted plots. POSIXct and POSIXt
  dates are currently not supported.  Package contents are
  geared towards economics and finance data
}
\author{
  Chandler Lutz \email{cl.eco@cbs.dk}
}
\keyword{package}

