% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recession_dates}
\alias{recession_dates}
\title{NBER Recession dates}
\format{A data frame with 2 columns:
\describe{
   \item{start}{Start of recession}
   \item{end}{End of recession}
}}
\source{
Chandler Lutz's calculations
}
\usage{
recession_dates
}
\description{
Recession dates from the NBER
}
\keyword{datasets}

