\name{std}
\alias{std}
\title{Standardize a vector to have mean zero and unit variance}
\usage{
  std(x)
}
\arguments{
  \item{x}{a vector to be standardized}
}
\value{
  a standardized vector
}
\description{
  Standardize a vector to have mean zero and unit variance
}
\author{
  Chandler Lutz \email{cl.eco@cbs.dk}
}

