% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidy_ts}
\alias{tidy_ts}
\title{Tidy up time series data for use with ggts() and ggts_facet()}
\usage{
tidy_ts(.data, ..., ts.name = NULL)
}
\arguments{
\item{.data}{the data (ts, mts, xts, zoo) which will be tidied}

\item{...}{variables to be passed to the dplyr select statement}

\item{ts.name}{(string) the name of the dataset. Necessary if .data
is a ts object}
}
\value{
(dataframe) a tidy dataframe with the time series data
}
\description{
Tidy up time series data for use with ggts() and ggts_facet()
}

