\docType{data}
\name{bear_dates}
\alias{bear_dates}
\title{Bear Market dates defined as a 20 percent or more
drop in the S&P500 over a period of two or more months
The first column is the start of the bear market;
the second column is the end of the bear market}
\description{
  Bear Market dates defined as a 20 percent or more drop in
  the S&P500 over a period of two or more months The first
  column is the start of the bear market; the second column
  is the end of the bear market
}
\keyword{datasets}

