% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggts.R
\name{ggts_facet}
\alias{ggts_facet}
\title{Plot a faceted time series -- multiple time series panels}
\usage{
ggts_facet(.data, ...)
}
\arguments{
\item{.data}{the data to be plotted. Either a dataframe with a time
column (e.g. a "Date", "POSIXlt", or "POSIXct" object), a ts or
mts object, zoo object, or xts object.}

\item{...}{arguments to be passed to dplyr select. If a dplyr
helper frunction is used, e.g. \code{matches}, dplyr must be
loaded or the function must be pasted via
\code{dplyr::matches("x")}}
}
\value{
a \code{ggplot} with the faceted time series
}
\description{
Plot a faceted time series -- multiple time series panels
}
\examples{
data(economis)
ggts_facet(economics)
ggts_facet(economics, psavert, uempmed)

library(xts)
data(AAPL.data) ##xts object
ggts_facet(AAPL.data)
ggts_facet(AAPL.data, -dplyr::matches("Volume"))

library(AER)
data(USMacroSW) ##mts object
ggts_facet(USMacroSW)

}

